food_safety_rag/requirements.txt
transformers
sentence-transformers
torch
datasets
pypdf
python-docx
python-pptx
elasticsearch
tqdm
python-dotenv